perbaikii error fitur lupa password (verifikasi email + no. handphone,
erorrawr

+++ database/migrations/2026_06_10_154033_create_transactions_table.php.php
@@ -8,32 +8,35 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('transactions', function (Blueprint $table) {
+        if (!Schema::hasTable('transactions')) {
+            Schema::create('transactions', function (Blueprint $table) {
 
-            $table->id();
+                $table->id();
 
-            $table->foreignId('user_id')
-                ->constrained()
-                ->onDelete('cascade');
+                $table->foreignId('user_id')
+                    ->constrained()
+                    ->onDelete('cascade');
 
-            $table->unsignedBigInteger('event_id');
+                $table->unsignedBigInteger('event_id');
 
-            $table->foreign('event_id')
-                ->references('id_event')
-                ->on('events')
-                ->onDelete('cascade');
+                $table->foreign('event_id')
+                    ->references('id_event')
+                    ->on('events')
+                    ->onDelete('cascade');
 
-            $table->string('ticket_type')->nullable();
-            $table->integer('qty');
-            $table->bigInteger('total_price');
+                $table->string('ticket_type')->nullable();
+                $table->integer('qty');
+                $table->bigInteger('total_price');
 
-            $table->string('payment_method');
-            $table->string('payment_proof')->nullable();
+                $table->string('payment_method');
+                $table->string('payment_proof')->nullable();
 
-            $table->string('status')->default('pending');
+                $table->string('status')->default('pending');
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
+        
     }
 
     public function down(): void
